simplify angular modules
