+++ plugins/CoreHome/angular/src/CoreHome.ts
@@ -2,4 +2,4 @@
  * Public API Surface of library
  */
 
-export * from './lib/library.module';
+export * from './lib/core-home.module';
