+++ plugins/CoreHome/angular/src/lib/core-home.module.ts
@@ -1,0 +1,21 @@
+import {NgModule, StaticProvider} from '@angular/core';
+import {downgradeModule, UpgradeModule} from '@angular/upgrade/static';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import {BrowserModule} from "@angular/platform-browser";
+
+@NgModule({
+  imports: [
+    UpgradeModule,
+    BrowserModule,
+  ],
+})
+export class CoreHomeModule {
+  ngDoBootstrap() {
+    // empty
+  }
+}
+
+const ng2BootstrapFn = (extraProviders: StaticProvider[]) =>
+    platformBrowserDynamic(extraProviders).bootstrapModule(CoreHomeModule);
+
+export const angularModuleName = downgradeModule(ng2BootstrapFn);
