--- plugins/CoreHome/angular/src/lib/library.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { LibraryService } from './library.service';
-
-describe('LibraryService', () => {
-  let service: LibraryService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(LibraryService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
