+++ plugins/ExampleAngular/angular/src/ExampleAngular.ts
@@ -2,4 +2,5 @@
  * Public API Surface of library
  */
 
-export * from './lib/library.module';
+export * from './lib/example.component';
+export * from './lib/example-angular.module';
