+++ plugins/ExampleAngular/angular/src/lib/example.component.ts
@@ -1,0 +1,25 @@
+declare var angular: angular.IAngularStatic;
+
+import {Component, OnInit, Type} from '@angular/core';
+import {downgradeComponent, UpgradeModule} from '@angular/upgrade/static';
+
+@Component({
+    selector: 'example-component',
+    template: `
+    <p>
+      this is a simple example angular component
+    </p>
+  `,
+    styles: [
+    ]
+})
+export class ExampleComponent implements OnInit {
+
+    constructor() { }
+
+    ngOnInit(): void {
+
+    }
+
+}
+
