--- plugins/ExampleAngular/angular/src/lib/library.component.ts
@@ -1,5 +1,0 @@
-declare var angular: angular.IAngularStatic;
-
-import {Component, OnInit, Type} from '@angular/core';
-import {downgradeComponent, UpgradeModule} from '@angular/upgrade/static';
-
