+++ plugins/Morpheus/templates/layout.twig
@@ -37,7 +37,7 @@
 
         {% include "_iframeBuster.twig" %}
         {% include "@CoreHome/_javaScriptDisabled.twig" %}
-<test-component></test-component>
+
         <div id="root">
             {% block root %}
             {% endblock %}
